+ asynchronous retrieving browser data + settings in statistics window added + allow to switch between linear and logarithmic plots + formatting of large numbers fixed

+++ source/Base/StringHelper.cpp
@@ -35,7 +35,7 @@ std::string StringHelper::format(float v, int fracPartDecimals)
     }
     while (fracPartDecimals-- > 0) {
         v *= 10;
-        result += std::to_string(static_cast<int>(v) % 10);
+        result += std::to_string(static_cast<uint64_t>(v) % 10);
     }
     return result;
 }
